Added oapiSaveSurface() function
