+++ Orbitersdk/include/OrbiterAPI.h
@@ -467,6 +467,19 @@ enum MatProp {
 };
 
 /**
+ * \brief Format specifiers for image files
+ */
+namespace oapi {
+	enum ImageFileFormat {
+		IMAGE_BMP = 0,
+		IMAGE_PNG = 1,
+		IMAGE_JPG = 2,
+		IMAGE_TIF = 3
+	};
+}
+
+
+/**
  * \brief Kepler orbital elements
  *
  * A set of 6 scalar parameters defining the state of an object in a 2-body
@@ -5504,6 +5517,18 @@ OAPIFUNC SURFHANDLE oapiCreateSurfaceEx (int width, int height, DWORD attrib);
 	* \return Surface handle for the loaded texture, or NULL if not found.
 	*/
 OAPIFUNC SURFHANDLE oapiLoadSurfaceEx(const char* fname, DWORD attrib, bool bPath = false);
+
+	/**
+	* \brief Save the contents of a surface to a formatted image file or to the clipboard
+	* \param hSrf surface handle (0 for primary render surface)
+	* \param fname image file path relative to orbiter root directory (excluding file extension), or NULL to save to clipboard
+	* \param fmt output file format
+	* \param quality quality request if the format supports it (0-1)
+	* \return Should return true on success
+	* \default Nothing, returns false
+	*/
+OAPIFUNC bool oapiSaveSurface(const char* fname, SURFHANDLE hSrf, oapi::ImageFileFormat fmt, float quality);
+
 
 	/**
 	* \brief Create a surface from a bitmap. Bitmap surfaces are typically used for blitting
